requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database - MySQL
mysqlclient==2.2.0  # This is the MySQL driver
# OR if mysqlclient fails to install, use:
# pymysql==1.1.0  # Pure Python MySQL client (slower but easier to install)

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.2
django-rest-auth==0.9.5

# Utilities
python-dotenv==1.0.0
django-environ==0.11.2
# pillow==9.5.0

# Payments
stripe==7.5.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.1

# Testing & Development
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Performance & Monitoring
django-debug-toolbar==4.2.0
django-silk==5.4.3

# File Storage
django-storages==1.14.2
boto3==1.34.0

# Caching
django-redis==5.4.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# Search
django-haystack==3.3.0
elasticsearch==8.11.1

# Rate Limiting
django-ratelimit==4.1.0

# Logging
python-json-logger==2.0.7

# Export/Import
# openpyxl==3.1.2
# pandas==2.1.4